requirements.txt
requirements:
pandas
pyyaml
requests
json